feat(dashboard): allow checkers to be copied
- Refactor CreateNewModal into its own component, and re-use
  in CheckerCard to facilitate duplication
- If a checker is supplied to CreateNewModal, make a copy of it,
  but clear out the id, title and description, forcing the user
  to re-enter those
